requirements.txt
# AI and ML Libraries
openai>=1.12.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-openai>=0.0.5

# API and Web
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0

# Data Processing
pandas>=2.2.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
